Added check function for designated sites
