+++ packages/web-service/src/pages/conservation-considerations/__tests__/designated-site-remove.spec.js
@@ -3,71 +3,6 @@ jest.spyOn(console, 'error').mockImplementation(() => null)
 describe('the special area remove functions', () => {
   beforeEach(() => jest.resetModules())
 
-  describe('the checkData function', () => {
-    it('returns null if there is no site to remove', async () => {
-      jest.doMock('../../../services/api-requests.js', () => ({
-        APIRequests: {
-          DESIGNATED_SITES: {
-            getDesignatedSites: jest.fn(() => [
-              {
-                id: 'fa5b8103-56a9-ed11-aad1-0022481b53bf',
-                siteName: 'Ribble Estuary',
-                siteType: 100000001
-              }
-            ]),
-            get: jest.fn(() => [])
-          }
-        }
-      }))
-      const request = {
-        query: { id: '344be97d-c928-4753-ae09-f8944ad9f228' },
-        cache: () => ({
-          getData: () => ({
-            applicationId: '26a3e94f-2280-4ea5-ad72-920d53c110fc'
-          }),
-          setData: jest.fn()
-        })
-      }
-      const { checkDesignatedSite } = await import('../designated-site-remove.js')
-      const h = { redirect: jest.fn() }
-      await checkDesignatedSite(request, h)
-      expect(h.redirect).toHaveBeenCalledWith('/designated-site-check-answers')
-    })
-
-    it('returns the check page if there is a site to remove', async () => {
-      jest.doMock('../../../services/api-requests.js', () => ({
-        APIRequests: {
-          DESIGNATED_SITES: {
-            getDesignatedSites: jest.fn(() => [
-              {
-                id: 'fa5b8103-56a9-ed11-aad1-0022481b53bf',
-                siteName: 'Ribble Estuary',
-                siteType: 100000001
-              }
-            ]),
-            get: jest.fn(() => [{
-              id: '344be97d-c928-4753-ae09-f8944ad9f228',
-              designatedSiteId: 'fa5b8103-56a9-ed11-aad1-0022481b53bf'
-            }])
-          }
-        }
-      }))
-      const request = {
-        query: { id: '344be97d-c928-4753-ae09-f8944ad9f228' },
-        cache: () => ({
-          getData: () => ({
-            applicationId: '26a3e94f-2280-4ea5-ad72-920d53c110fc'
-          }),
-          setData: jest.fn()
-        })
-      }
-      const { checkDesignatedSite } = await import('../designated-site-remove.js')
-      const h = { redirect: jest.fn() }
-      await checkDesignatedSite(request, h)
-      expect(h.redirect).not.toHaveBeenCalled()
-    })
-  })
-
   describe('the getData function', () => {
     it('returns the site name to be deleted', async () => {
       const mockSet = jest.fn()
