+++ packages/web-service/src/pages/conservation-considerations/details-of-permission.js
@@ -2,7 +2,7 @@ import Joi from 'joi'
 import pageRoute from '../../routes/page-route.js'
 import { conservationConsiderationURIs } from '../../uris.js'
 import { checkApplication } from '../common/check-application.js'
-import { allCompletion, getCurrentSite } from './common.js'
+import { allCompletion, checkDesignatedSite, getCurrentSite } from './common.js'
 import { APIRequests } from '../../services/api-requests.js'
 
 const { OWNER_PERMISSION_DETAILS } = conservationConsiderationURIs
@@ -22,7 +22,7 @@ export const setData = async request => {
 export default pageRoute({
   page: OWNER_PERMISSION_DETAILS.page,
   uri: OWNER_PERMISSION_DETAILS.uri,
-  checkData: checkApplication,
+  checkData: [checkApplication, checkDesignatedSite],
   getData: getData,
   validator: Joi.object({
     'permission-details': Joi.string().replace('\r\n', '\n').trim().required().max(4000)
